connect to taskserver, WIP RSA DER parsing
The RSA DER parsing doesn't seem to be necessary for UWP, only on .NET, and will likely be removed in a subsequent commit.

The certificate is currently taken from the system store, where the user has to install it previously. This works, but is discouraged to the degree that apps doing it are not accepted on the Windows Store. It will need to be done differently.